services/ocr-service/requirements.txt
paddleocr==3.3.2
opencv-python
python-multipart
pillow
grpcio
grpcio-tools
numpy